requirements-vision.txt
flask
flask-cors
opencv-python
torch
torchvision
pillow
imagehash
aiohttp
python-dotenv
numpy
